fix: update size_t to int32_t

+++ src/mouse.c
@@ -132,8 +132,8 @@ void moveMouse(MMSignedPoint point)
 	//Mouse motion is now done using SendInput with MOUSEINPUT. We use Absolute mouse positioning
 	#define MOUSE_COORD_TO_ABS(coord, width_or_height) ((65536 * (coord) / width_or_height) + ((coord) < 0 ? -1 : 1))
 
-	size_t x = MOUSE_COORD_TO_ABS(point.x-vscreenMinX, vscreenWidth);
-	size_t y = MOUSE_COORD_TO_ABS(point.y-vscreenMinY, vscreenHeight);
+	int32_t x = MOUSE_COORD_TO_ABS(point.x-vscreenMinX, vscreenWidth);
+	int32_t y = MOUSE_COORD_TO_ABS(point.y-vscreenMinY, vscreenHeight);
 
 	INPUT mouseInput = {0};
 	mouseInput.type = INPUT_MOUSE;
